+++ controllers/QrController.php
@@ -139,7 +139,7 @@ class QrController extends Controller
                 ]
                 )
                 ->exists();
-            if (!$exist) {
+            if (!$exist && Yii::$app->user->can('organisatieOpstart')) {
                 $model->delete();
                 Yii::$app->session->setFlash('success', Yii::t('app', 'Deleted silent station.'));
             } else {
