+++ controllers/RouteController.php
@@ -203,7 +203,7 @@ class RouteController extends Controller
                     ->exists();
             }
 
-            if (!$exist) {
+            if (!$exist && Yii::$app->user->can('organisatieOpstart')) {
                 $model->delete();
                 Yii::$app->session->setFlash('success', Yii::t('app', 'Deleted route.'));
             } else {
